% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_template_path}
\alias{get_template_path}
\title{Helper to resolve template paths across environments}
\usage{
get_template_path(template_name)
}
\description{
Helper to resolve template paths across environments
}
\keyword{internal}
