% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{plot_mcmc_acf}
\alias{plot_mcmc_acf}
\title{Generate Autocorrelograms (ACF) for MCMC Chains}
\usage{
plot_mcmc_acf(samples, param_name)
}
\arguments{
\item{samples}{A numeric vector or matrix of MCMC samples.}

\item{param_name}{Name of the parameter for labeling the plot.}
}
\value{
A ggplot object.
}
\description{
Generate Autocorrelograms (ACF) for MCMC Chains
}
