% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{plot_mcmc_trace}
\alias{plot_mcmc_trace}
\title{Generate Publication-Quality Trace Plots}
\usage{
plot_mcmc_trace(samples, param_name)
}
\arguments{
\item{samples}{A matrix or data frame of MCMC samples.}

\item{param_name}{Name of the parameter for labeling the plot.}
}
\description{
Generate Publication-Quality Trace Plots
}
