% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{project_config}
\alias{project_config}
\title{Internal Project Constants and Path Configuration}
\format{
An object of class \code{list} of length 21.
}
\usage{
project_config
}
\description{
This file provides a centralized location for directory paths and project-wide
constants. Similar to a 'constants.py' in Python, this ensures consistency
and easier maintenance across the MCMC engine.
}
\keyword{internal}
