% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rinvgamma}
\alias{rinvgamma}
\title{Sample from an Inverse-Gamma Distribution}
\usage{
rinvgamma(n, shape, scale)
}
\arguments{
\item{n}{Number of samples.}

\item{shape}{Shape parameter (a).}

\item{scale}{Scale parameter (b).}
}
\value{
A numeric vector of inverse-gamma samples.
}
\description{
This is used for sampling the variance parameter \(\sigma^2\) in the Gibbs sampler.
}
