% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_engine.R
\name{run_gibbs_sampler}
\alias{run_gibbs_sampler}
\title{Bayesian MCMC AR(p) Gibbs Sampler}
\usage{
run_gibbs_sampler(
  y,
  p,
  n_iter = project_config$default_n_iter,
  burn_in = project_config$default_burn_in,
  priors = NULL
)
}
\arguments{
\item{y}{Numeric vector of the observed time series.}

\item{p}{Integer; the order of the autoregressive process.}

\item{n_iter}{Total number of MCMC iterations.}

\item{burn_in}{Number of initial iterations to discard.}

\item{priors}{A list containing hyperparameters for the MVN and IG priors.}
}
\value{
A list containing the posterior samples for \(\beta\) and \(\sigma^2\).
}
\description{
Fits an AR(p) model using a Bayesian Gibbs sampling approach with conjugate priors.
}
