% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{summarize_mcmc_results}
\alias{summarize_mcmc_results}
\title{Calculate MCMC Summary Statistics and Effective Sample Size}
\usage{
summarize_mcmc_results(samples)
}
\arguments{
\item{samples}{A numeric matrix or data frame of MCMC samples.}
}
\value{
A data frame containing Mean, SD, Quatiles, and ESS.
}
\description{
Calculate MCMC Summary Statistics and Effective Sample Size
}
